Refactor user forms
Refactor Settings, EditUser, NewUser to use a shared form
